api/requirements.txt
# this is needen because ena-upload is depending on the PyYAML>5.* 
# but version 5.4.1 cannot be installed because of a malicious setup.cfg
PyYAML==5.3.1
ena-upload-cli==0.6.3
click==8.1.7
python-box==7.1.1
rich==13.7.0
pandas==2.1.3
django==4.2.7
django_extensions==3.2.3
django-cors-headers==4.3.1
django-filter==23.3
django-dbbackup==4.0.2
djangorestframework==3.14.0
djangorestframework-csv==3.0.1
drf-spectacular[sidecar]==0.26.5
drf-schema-adapter==3.0.6
gunicorn==21.2.0
uvicorn==0.24.0.post1
psycopg2==2.9.9
sh==2.0.6
